+++ packages/desktop/lib/electron/apis/ledger.api.ts
@@ -1,13 +1,8 @@
 import { ipcRenderer } from 'electron'
-import { LedgerMethod } from '../enums/ledger-method.enum'
+import { LedgerApiMethod } from '@core/ledger/enums'
 
 export default {
-    makeRequest: (method: LedgerMethod, ...parameters: unknown[]) => {
+    makeRequest: (method: LedgerApiMethod, ...parameters: unknown[]) => {
         ipcRenderer.send(method, ...parameters)
     }
-}
-
-export interface ILedgerApiBridge {
-    makeRequest(method: LedgerMethod.GenerateEvmAddress, bip32Path: string, verify: boolean): void
-    makeRequest(method: LedgerMethod.SignEvmTransaction, transactionHex: string, bip32Path: string): void
 }
