+++ packages/shared/src/lib/core/ledger/constants/default-ledger-api-request-configuration.constant.ts
@@ -1,0 +1,6 @@
+import { ILedgerApiRequestOptions } from '../interfaces'
+
+export const DEFAULT_LEDGER_API_REQUEST_CONFIGURATION: ILedgerApiRequestOptions = {
+    timeout: 60,
+    pollingInterval: 100,
+}
