+++ packages/shared/src/lib/core/ledger/constants/index.ts
@@ -1,3 +1,4 @@
+export * from './default-ledger-api-request-configuration.constant'
 export * from './default-ledger-nano-status-poll-interval.constant'
 export * from './ledger-error-locales.constant'
 export * from './use-ledger-simulator.constant'
