+++ packages/shared/src/lib/core/ledger/enums/index.ts
@@ -1,2 +1,3 @@
+export * from './ledger-api-method.enum'
 export * from './ledger-app-name.enum'
 export * from './ledger-error.enum'
