+++ packages/shared/src/lib/core/ledger/enums/ledger-api-method.enum.ts
@@ -1,4 +1,4 @@
-export enum LedgerMethod {
+export enum LedgerApiMethod {
     GenerateEvmAddress = 'generate-evm-address',
     SignEvmTransaction = 'sign-evm-transaction',
 }
