+++ packages/shared/src/lib/core/ledger/index.ts
@@ -4,4 +4,5 @@ export * from './constants'
 export * from './enums'
 export * from './interfaces'
 export * from './stores'
+export * from './types'
 export * from './utils'
