+++ packages/shared/src/lib/core/ledger/interfaces/index.ts
@@ -1,3 +1,5 @@
-export * from './ledger.interface'
+export * from './ledger-api.interface'
+export * from './ledger-api-bridge.interface'
+export * from './ledger-api-request-options.interface'
 export * from './ledger-connection-state.interface'
 export * from './ledger-nano-status-polling-configuration.interface'
