+++ packages/shared/src/lib/core/ledger/interfaces/ledger-api-request-options.interface.ts
@@ -1,0 +1,4 @@
+export interface ILedgerApiRequestOptions {
+    timeout: number
+    pollingInterval: number
+}
