+++ packages/shared/src/lib/core/ledger/interfaces/ledger-api.interface.ts
@@ -1,6 +1,6 @@
 import { EvmTransactionData } from '@core/layer-2/types'
 
-export interface ILedger {
+export interface ILedgerApi {
     generateEvmAddress(accountIndex: number, coinType: number, verify?: boolean): Promise<string>
     signEvmTransaction(transactionData: EvmTransactionData, bip32Path: string): Promise<string>
 }
