+++ packages/shared/src/lib/core/ledger/types/ledger-api-request-response.type.ts
@@ -1,0 +1,4 @@
+import { IEvmAddress } from '@core/app/interfaces'
+import { IEvmTransactionSignature } from '@core/layer-2/interfaces'
+
+export type LedgerApiRequestResponse = IEvmAddress | IEvmTransactionSignature
